Project/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
decouple==0.0.7
Django==4.2.2
django-filter==23.2
django-recaptcha==3.0.0
django-recaptcha3==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3

pip~=22.3.1
config~=0.5.1
setuptools~=65.5.0
django-debug-toolbar~=4.1.0
django-modeltranslation~=0.18.10